MongoDb is ready to be in use
