+++ build.gradle
@@ -19,6 +19,7 @@ dependencies {
 	compile group: 'org.eclipse.paho', name: 'org.eclipse.paho.client.mqttv3', version: '1.2.1'
 	compileOnly 'org.projectlombok:lombok:1.18.6'
 	compile group: 'org.json', name: 'json', version: '20160810'
+	compile("org.springframework.boot:spring-boot-starter-data-mongodb")
 
 	annotationProcessor 'org.projectlombok:lombok:1.18.6'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
