+++ src/main/java/fi/robosailboat/webservice/boatCommunication/dto/Command.java
@@ -1,4 +1,4 @@
-package fi.robosailboat.webservice.boatCommunication;
+package fi.robosailboat.webservice.boatCommunication.dto;
 
 import lombok.Data;
