+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
-
+spring.data.mongodb.uri=mongodb+srv://RoboSailboat:RoboSailBoat@robosailboat-zrbyz.mongodb.net/test?retryWrites=true
+spring.data.mongodb.database=RoboSailboat
